+++ sdk/cosmos/azure_data_cosmos/tests/emulator_tests/cosmos_feed_ranges.rs
@@ -5,10 +5,8 @@ use super::framework;
 
 use std::error::Error;
 
-use azure_data_cosmos::{
-    models::{ContainerProperties, ThroughputProperties},
-    CreateContainerOptions,
-};
+use azure_data_cosmos::models::{ContainerProperties, ThroughputProperties};
+use azure_data_cosmos::options::CreateContainerOptions;
 use base64::Engine;
 
 use framework::TestClient;