+++ sdk/cosmos/azure_data_cosmos/tests/emulator_tests/cosmos_response_metadata.rs
@@ -5,9 +5,10 @@
 use super::framework;
 
 use azure_core::{http::StatusCode, Uuid};
-use azure_data_cosmos::models::{ContainerProperties, ResponseHeaders};
+use azure_data_cosmos::models::ContainerProperties;
+use azure_data_cosmos::models::ResponseHeaders;
 use azure_data_cosmos::Query;
-use azure_data_cosmos::{clients::ContainerClient, query::FeedScope};
+use azure_data_cosmos::{clients::ContainerClient, feed::FeedScope};
 use framework::{TestClient, TestRunContext};
 use futures::StreamExt;
 use serde::{Deserialize, Serialize};